cbvccm/cbvultimate/requirements.txt
Django==1.9.7
django-appconf==1.0.2
django-avatar==3.0.0
django-countries==3.4.1
django-country==0.1.0
django-shortuuidfield==0.1.3
Pillow==3.3.0
psycopg2==2.6.1
shortuuid==0.4.3
six==1.10.0
